Finished and applied patch #5638: setWindSpeed
3 new (client-side) functions:

1. setWindSpeed ( float speedX, float speedY, float speedZ )

Makes trees go whee, depending on the values

2. getWindSpeed ( )

Returns the whee-ness of trees

3. resetWindSpeed ( )

Changes the whee-ness back to default

Server-side stuff should probably be made as well, but a small conversation about video cards got in the way. So I should look back into that at a later moment

Credits to onestone, for creating the original patch, x86 who slapped me around a bit with the most obvious mistake in history, and myself for, well, everything else.